refactored copy and added types, removed limit
