+++ src/containers/announcements/ducks/thunks.ts
@@ -1,15 +1,13 @@
 import { Announcement, AnnouncementsThunkAction } from './types';
 import { announcements } from './actions';
 
-export const getAnnouncements = (
-  limit?: number,
-): AnnouncementsThunkAction<void> => {
+export const getAnnouncements = (): AnnouncementsThunkAction<void> => {
   return (dispatch, getState, { publicApiClient }) => {
     dispatch(announcements.loading());
     return publicApiClient
       .getAnnouncements()
       .then((response: Announcement[]) => {
-        return limit && limit >= 0 ? response.slice(0, limit) : response;
+        return response;
       })
       .then((response: Announcement[]) => {
         dispatch(announcements.loaded(response));
