+++ src/utils/copy.tsx
@@ -1,0 +1,2 @@
+export const NO_ANNOUNCEMENTS_HEADER = 'There are currently no announcements!';
+export const NO_ANNOUNCEMENTS_SUBHEADER = 'Come back later for future updates!';
